requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic-settings
celery
redis
boto3
pytesseract
pillow
httpx
python-dotenv
pytest
pytest-asyncio
httpx
